+++ test/absinthe/type/enum_test.exs
@@ -2,6 +2,7 @@ defmodule Absinthe.Type.EnumTest do
   use Absinthe.Case, async: true
 
   alias Absinthe.Type
+  alias Absinthe.Fixtures.Enums
 
   defmodule TestSchema do
     use Absinthe.Schema
@@ -50,190 +51,6 @@ defmodule Absinthe.Type.EnumTest do
     end
   end
 
-  defmodule TestSchemaEnumValueDescriptionKeyword do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    enum :description_keyword_argument do
-      value :normal_string, description: "string"
-      value :local_function_call, description: test_function("red")
-
-      value :function_call_using_absolute_path_to_current_module,
-        description:
-          Absinthe.Type.EnumTest.TestSchemaEnumValueDescriptionKeyword.test_function("red")
-
-      value :standard_library_function, description: String.replace("red", "e", "a")
-      value :function_in_nested_module, description: NestedModule.nested_function("hello")
-
-      value :external_module_function_call,
-        description: Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-
-      value :module_attribute_string_concat, description: "hello " <> @module_attribute
-      value :interpolation_of_module_attribute, description: "hello #{@module_attribute}"
-    end
-  end
-
-  defmodule TestSchemaEnumDescriptionKeyword do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    enum :normal_string, description: "string" do
-    end
-
-    enum :local_function_call, description: test_function("red") do
-    end
-
-    enum :function_call_using_absolute_path_to_current_module,
-      description: Absinthe.Type.EnumTest.TestSchemaEnumDescriptionKeyword.test_function("red") do
-    end
-
-    enum :standard_library_function, description: String.replace("red", "e", "a") do
-    end
-
-    enum :function_in_nested_module, description: NestedModule.nested_function("hello") do
-    end
-
-    enum :external_module_function_call,
-      description: Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello") do
-    end
-
-    enum :module_attribute_string_concat, description: "hello " <> @module_attribute do
-    end
-
-    enum :interpolation_of_module_attribute, description: "hello #{@module_attribute}" do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-  end
-
-  defmodule TestSchemaEnumDescriptionAttribute do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    @desc "string"
-    enum :normal_string do
-    end
-
-    # These tests do not work as test_function is not available at compile time, and the
-    # expression for the @desc attribute is evaluated at compile time. There is nothing we can
-    # really do about it
-
-    # @desc test_function("red")
-    # enum :local_function_call do
-    # end
-
-    # @desc Absinthe.Type.EnumTest.TestSchemaEnumAttribute.test_function("red")
-    # enum :function_call_using_absolute_path_to_current_module do
-    # end
-
-    @desc String.replace("red", "e", "a")
-    enum :standard_library_function do
-    end
-
-    @desc NestedModule.nested_function("hello")
-    enum :function_in_nested_module do
-    end
-
-    @desc Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-    enum :external_module_function_call do
-    end
-
-    @desc "hello " <> @module_attribute
-    enum :module_attribute_string_concat do
-    end
-
-    @desc "hello #{@module_attribute}"
-    enum :interpolation_of_module_attribute do
-    end
-  end
-
-  defmodule TestSchemaEnumDescriptionMacro do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    enum :normal_string do
-      description "string"
-    end
-
-    enum :local_function_call do
-      description test_function("red")
-    end
-
-    enum :function_call_using_absolute_path_to_current_module do
-      description Absinthe.Type.EnumTest.TestSchemaEnumDescriptionMacro.test_function("red")
-    end
-
-    enum :standard_library_function do
-      description String.replace("red", "e", "a")
-    end
-
-    enum :function_in_nested_module do
-      description NestedModule.nested_function("hello")
-    end
-
-    enum :external_module_function_call do
-      description Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-    end
-
-    enum :module_attribute_string_concat do
-      description "hello " <> @module_attribute
-    end
-
-    enum :interpolation_of_module_attribute do
-      description "hello #{@module_attribute}"
-    end
-  end
-
   describe "enums" do
     test "can be defined by a map with defined values" do
       type = TestSchema.__absinthe_type__(:color_channel)
@@ -264,7 +81,7 @@ defmodule Absinthe.Type.EnumTest do
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
         type =
-          TestSchemaEnumValueDescriptionKeyword.__absinthe_type__(:description_keyword_argument)
+          Enums.TestSchemaEnumValueDescriptionKeyword.__absinthe_type__(:description_keyword_argument)
 
         assert type.values[unquote(test_label)].description == unquote(expected_value)
       end
@@ -278,7 +95,7 @@ defmodule Absinthe.Type.EnumTest do
                       expected_value: expected_value
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
-        type = TestSchemaEnumDescriptionKeyword.__absinthe_type__(unquote(test_label))
+        type = Enums.TestSchemaEnumDescriptionKeyword.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
@@ -300,7 +117,7 @@ defmodule Absinthe.Type.EnumTest do
                       expected_value: expected_value
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
-        type = TestSchemaEnumDescriptionAttribute.__absinthe_type__(unquote(test_label))
+        type = Enums.TestSchemaEnumDescriptionAttribute.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
@@ -313,7 +130,7 @@ defmodule Absinthe.Type.EnumTest do
                       expected_value: expected_value
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
-        type = TestSchemaEnumDescriptionMacro.__absinthe_type__(unquote(test_label))
+        type = Enums.TestSchemaEnumDescriptionMacro.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
