+++ test/absinthe/type/input_object_test.exs
@@ -1,6 +1,8 @@
 defmodule Absinthe.Type.InputObjectTest do
   use Absinthe.Case, async: true
 
+  alias Absinthe.Fixtures.InputObject
+
   defmodule Schema do
     use Absinthe.Schema
 
@@ -12,265 +14,6 @@ defmodule Absinthe.Type.InputObjectTest do
     input_object :profile do
       field :name, :string
       field :profile_picture, :string
-    end
-  end
-
-  defmodule TestSchemaInputObjectDescriptionKeyword do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    input_object :normal_string, description: "string" do
-    end
-
-    input_object :local_function_call, description: test_function("red") do
-    end
-
-    input_object :function_call_using_absolute_path_to_current_module,
-      description:
-        Absinthe.Type.InputObjectTest.TestSchemaInputObjectDescriptionKeyword.test_function("red") do
-    end
-
-    input_object :standard_library_function, description: String.replace("red", "e", "a") do
-    end
-
-    input_object :function_in_nested_module, description: NestedModule.nested_function("hello") do
-    end
-
-    input_object :external_module_function_call,
-      description: Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello") do
-    end
-
-    input_object :module_attribute_string_concat, description: "hello " <> @module_attribute do
-    end
-
-    input_object :interpolation_of_module_attribute, description: "hello #{@module_attribute}" do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-  end
-
-  defmodule TestSchemaInputObjectDescriptionAttribute do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    @desc "string"
-    input_object :normal_string do
-    end
-
-    # These tests do not work as test_function is not available at compile time, and the
-    # expression for the @desc attribute is evaluated at compile time. There is nothing we can
-    # really do about it
-
-    # @desc test_function("red")
-    # input_object :local_function_call do
-    # end
-
-    # @desc Absinthe.Type.InputObjectTest.TestSchemaInputObjectAttribute.test_function("red")
-    # input_object :function_call_using_absolute_path_to_current_module do
-    # end
-
-    @desc String.replace("red", "e", "a")
-    input_object :standard_library_function do
-    end
-
-    @desc NestedModule.nested_function("hello")
-    input_object :function_in_nested_module do
-    end
-
-    @desc Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-    input_object :external_module_function_call do
-    end
-
-    @desc "hello " <> @module_attribute
-    input_object :module_attribute_string_concat do
-    end
-
-    @desc "hello #{@module_attribute}"
-    input_object :interpolation_of_module_attribute do
-    end
-  end
-
-  defmodule TestSchemaInputObjectDescriptionMacro do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    input_object :normal_string do
-      description "string"
-    end
-
-    input_object :local_function_call do
-      description test_function("red")
-    end
-
-    input_object :function_call_using_absolute_path_to_current_module do
-      description Absinthe.Type.InputObjectTest.TestSchemaInputObjectDescriptionMacro.test_function(
-                    "red"
-                  )
-    end
-
-    input_object :standard_library_function do
-      description String.replace("red", "e", "a")
-    end
-
-    input_object :function_in_nested_module do
-      description NestedModule.nested_function("hello")
-    end
-
-    input_object :external_module_function_call do
-      description Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-    end
-
-    input_object :module_attribute_string_concat do
-      description "hello " <> @module_attribute
-    end
-
-    input_object :interpolation_of_module_attribute do
-      description "hello #{@module_attribute}"
-    end
-  end
-
-  defmodule TestSchemaInputObjectFieldsAndArgsDescription do
-    use Absinthe.Schema
-    @module_attribute "goodbye"
-
-    defmodule NestedModule do
-      def nested_function(arg1) do
-        arg1
-      end
-    end
-
-    query do
-    end
-
-    def test_function(arg1) do
-      arg1
-    end
-
-    input_object :description_keyword_argument do
-      field :normal_string, :string, description: "string"
-      field :local_function_call, :string, description: test_function("red")
-
-      field :function_call_using_absolute_path_to_current_module, :string,
-        description:
-          Absinthe.Type.InputObjectTest.TestSchemaInputObjectFieldsAndArgsDescription.test_function(
-            "red"
-          )
-
-      field :standard_library_function, :string, description: String.replace("red", "e", "a")
-
-      field :function_in_nested_module, :string,
-        description: NestedModule.nested_function("hello")
-
-      field :external_module_function_call, :string,
-        description: Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-
-      field :module_attribute_string_concat, :string, description: "hello " <> @module_attribute
-      field :interpolation_of_module_attribute, :string, description: "hello #{@module_attribute}"
-    end
-
-    input_object :description_attribute do
-      @desc "string"
-      field :normal_string, :string
-
-      # These tests do not work as test_function is not available at compile time, and the
-      # expression for the @desc attribute is evaluated at compile time. There is nothing we can
-      # really do about it
-
-      # @desc test_function("red")
-      # field :local_function_call, :string
-
-      # @desc Absinthe.Type.InputObjectTest.TestSchemaInputObjectFieldsAndArgsDescription.test_function(
-      #         "red"
-      #       )
-      # field :function_call_using_absolute_path_to_current_module, :string
-
-      @desc String.replace("red", "e", "a")
-      field :standard_library_function, :string
-
-      @desc NestedModule.nested_function("hello")
-      field :function_in_nested_module, :string
-
-      @desc Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-      field :external_module_function_call, :string
-
-      @desc "hello " <> @module_attribute
-      field :module_attribute_string_concat, :string
-
-      @desc "hello #{@module_attribute}"
-      field :interpolation_of_module_attribute, :string
-    end
-
-    input_object :field_description_macro do
-      field :normal_string, :string do
-        description "string"
-      end
-
-      field :local_function_call, :string do
-        description test_function("red")
-      end
-
-      field :function_call_using_absolute_path_to_current_module, :string do
-        description Absinthe.Type.InputObjectTest.TestSchemaInputObjectFieldsAndArgsDescription.test_function(
-                      "red"
-                    )
-      end
-
-      field :standard_library_function, :string do
-        description String.replace("red", "e", "a")
-      end
-
-      field :function_in_nested_module, :string do
-        description NestedModule.nested_function("hello")
-      end
-
-      field :external_module_function_call, :string do
-        description Absinthe.Fixtures.FunctionEvaluationHelpers.external_function("hello")
-      end
-
-      field :module_attribute_string_concat, :string do
-        description "hello " <> @module_attribute
-      end
-
-      field :interpolation_of_module_attribute, :string do
-        description "hello #{@module_attribute}"
-      end
     end
   end
 
@@ -295,7 +38,7 @@ defmodule Absinthe.Type.InputObjectTest do
                       expected_value: expected_value
                     } ->
       test "for `#{test_label}` (evaluates description to `'#{expected_value}'`)" do
-        type = TestSchemaInputObjectDescriptionKeyword.__absinthe_type__(unquote(test_label))
+        type = InputObject.TestSchemaInputObjectDescriptionKeyword.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
@@ -317,7 +60,7 @@ defmodule Absinthe.Type.InputObjectTest do
                       expected_value: expected_value
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
-        type = TestSchemaInputObjectDescriptionAttribute.__absinthe_type__(unquote(test_label))
+        type = InputObject.TestSchemaInputObjectDescriptionAttribute.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
@@ -330,7 +73,7 @@ defmodule Absinthe.Type.InputObjectTest do
                       expected_value: expected_value
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
-        type = TestSchemaInputObjectDescriptionMacro.__absinthe_type__(unquote(test_label))
+        type = InputObject.TestSchemaInputObjectDescriptionMacro.__absinthe_type__(unquote(test_label))
         assert type.description == unquote(expected_value)
       end
     end)
@@ -344,7 +87,7 @@ defmodule Absinthe.Type.InputObjectTest do
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
         type =
-          TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(
+          InputObject.TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(
             :description_keyword_argument
           )
 
@@ -370,7 +113,7 @@ defmodule Absinthe.Type.InputObjectTest do
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
         type =
-          TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(:description_attribute)
+          InputObject.TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(:description_attribute)
 
         assert type.fields[unquote(test_label)].description == unquote(expected_value)
       end
@@ -385,7 +128,7 @@ defmodule Absinthe.Type.InputObjectTest do
                     } ->
       test "for #{test_label} (evaluates description to '#{expected_value}')" do
         type =
-          TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(
+          InputObject.TestSchemaInputObjectFieldsAndArgsDescription.__absinthe_type__(
             :field_description_macro
           )
